// server/src/documents/enums/document-type.enum.ts
export enum DocumentType {
  PROTOCOL = 'PROTOCOL',
  INFORMED_CONSENT_FORM = 'INFORMED_CONSENT_FORM',
  PRINCIPAL_INVESTIGATOR_CV = 'PRINCIPAL_INVESTIGATOR_CV',
  ETHICS_TRAINING_CERT = 'ETHICS_TRAINING_CERT',
  NHRA_RESEARCHER_CERT = 'NHRA_RESEARCHER_CERT',
  COVER_LETTER = 'COVER_LETTER',
  BUDGET = 'BUDGET',
  OTHER = 'OTHER',
}